Revert "mokee-sdk self-removing prefs: move reap from onBindViewHolder to onAttached"
This reverts commit 027b615325092f7676a34c137248dc9fcd57cc35.

While it served its original purpose, this was causing bad side-effects
when using our constraints (verified at least with requiresConfig and
requiresAction). Moving the reap to onAttached was rendering self-removing
prefs useless because if the view isn't loaded yet, the preference
(which also doesn't yet exists) can't really be hidden.

Change-Id: Ic61fe4b9979412d89c55ea1543c494e69bf6eab7