mokee-sdk: avoid NPE caused by early interface refresh
When switching data SIM, there is a great chance where underlying
link is not ready on the arrival of CONNECTIVITY_ACTION broadcast.
getLinkProperties() will return null in such case and cause a SystemUI
crash.

In order to solve this, we only mark the interface stats as dirty in
the broadcast, and delay the actual work until next periodic display
refresh - link properties will be more likely to be ready then.

We also check the return values of method invocations. The logic will
retry if any link property is null, and no display update can happen
before all active interfaces are acquired.

Change-Id: Idf828a57c729c86959429abf1f23392e47c26e05